org: integrate qt tests in pytest
- Integrate QT tests to the pytest collector.
- Remove  ~export_values()~ --  it  was  not fully  clear  from the  pybind
  documentation,  but to  work with  standalone  enums this  option is  not
  necessary as it pushes the enum values to the upper scope, clobbering the
  name resolution.
