+++ .envrc
@@ -4,7 +4,7 @@ export PYTHON_KEYRING_BACKEND=keyring.backends.null.Keyring
 # export LLVM_LIBS_PATH="${LLVM_DIR}/lib/x86_64-unknown-linux-gnu"
 # export LD_LIBRARY_PATH="${LLVM_CLANG_LIBS_PATH}:${LLVM_LIBS_PATH}"
 # export PATH="$PWD/toolchain/llvm/bin:$PATH"
-# export LD_PRELOAD="$PWD/toolchain/llvm/lib/clang/17/lib/x86_64-unknown-linux-gnu/libclang_rt.asan.so"
+export LD_PRELOAD="$PWD/toolchain/llvm/lib/clang/17/lib/x86_64-unknown-linux-gnu/libclang_rt.asan.so"
 # export LD_PRELOAD="/usr/lib64/gcc/x86_64-pc-linux-gnu/13.2.1/../../../../lib64/libasan.so"
 # export ASAN_SYMBOLIZER_PATH="${LLVM_DIR}/bin/llvm-symbolizer"
 export ASAN_OPTIONS="detect_leaks=0"
