+++ src/py_libs/pyhaxorg/pyhaxorg.cpp
@@ -198,7 +198,6 @@ node can have subnodes.)RAW")
     .value("Inline", sem::Export::Format::Inline, R"RAW(Export directly in the paragraph)RAW")
     .value("Line", sem::Export::Format::Line, R"RAW(Single line of export)RAW")
     .value("Block", sem::Export::Format::Block, R"RAW(Multiple lines of export)RAW")
-    .export_values()
     ;
   pybind11::class_<sem::AdmonitionBlock, sem::SemId<sem::AdmonitionBlock>, sem::Block>(m, "AdmonitionBlock")
     .def_readwrite("loc", &sem::AdmonitionBlock::loc, R"RAW(Document)RAW")
@@ -252,18 +251,15 @@ node can have subnodes.)RAW")
     .value("RemoveCallout", sem::Code::Switch::Kind::RemoveCallout)
     .value("EmphasizeLine", sem::Code::Switch::Kind::EmphasizeLine)
     .value("Dedent", sem::Code::Switch::Kind::Dedent)
-    .export_values()
     ;
   pybind11::enum_<sem::Code::Results>(m, "CodeResults")
     .value("Replace", sem::Code::Results::Replace, R"RAW(Remove old result, replace with new value)RAW")
-    .export_values()
     ;
   pybind11::enum_<sem::Code::Exports>(m, "CodeExports")
     .value("None", sem::Code::Exports::None, R"RAW(Hide both original code and run result)RAW")
     .value("Both", sem::Code::Exports::Both, R"RAW(Show output and code)RAW")
     .value("Code", sem::Code::Exports::Code, R"RAW(Show only code)RAW")
     .value("Results", sem::Code::Exports::Results, R"RAW(Show only evaluation results)RAW")
-    .export_values()
     ;
   pybind11::class_<sem::Time, sem::SemId<sem::Time>, sem::Org>(m, "Time")
     .def_readwrite("loc", &sem::Time::loc, R"RAW(Document)RAW")
@@ -284,7 +280,6 @@ node can have subnodes.)RAW")
     .value("Exact", sem::Time::Repeat::Mode::Exact, R"RAW(?)RAW")
     .value("FirstMatch", sem::Time::Repeat::Mode::FirstMatch, R"RAW(Repeat on the first matching day in the future)RAW")
     .value("SameDay", sem::Time::Repeat::Mode::SameDay, R"RAW(Repeat task on the same day next week/month/year)RAW")
-    .export_values()
     ;
   pybind11::enum_<sem::Time::Repeat::Period>(m, "TimeRepeatPeriod")
     .value("Year", sem::Time::Repeat::Period::Year)
@@ -293,7 +288,6 @@ node can have subnodes.)RAW")
     .value("Day", sem::Time::Repeat::Period::Day)
     .value("Hour", sem::Time::Repeat::Period::Hour)
     .value("Minute", sem::Time::Repeat::Period::Minute)
-    .export_values()
     ;
   pybind11::class_<sem::Time::Static>(m, "TimeStatic")
     .def(pybind11::init<>())
@@ -307,7 +301,6 @@ node can have subnodes.)RAW")
   pybind11::enum_<sem::Time::TimeKind>(m, "TimeTimeKind")
     .value("Static", sem::Time::TimeKind::Static)
     .value("Dynamic", sem::Time::TimeKind::Dynamic)
-    .export_values()
     ;
   pybind11::class_<sem::TimeRange, sem::SemId<sem::TimeRange>, sem::Org>(m, "TimeRange")
     .def_readwrite("loc", &sem::TimeRange::loc, R"RAW(Document)RAW")
@@ -358,7 +351,6 @@ node can have subnodes.)RAW")
     .value("Added", sem::SubtreeLog::Priority::Action::Added)
     .value("Removed", sem::SubtreeLog::Priority::Action::Removed)
     .value("Changed", sem::SubtreeLog::Priority::Action::Changed)
-    .export_values()
     ;
   pybind11::class_<sem::SubtreeLog::Note>(m, "SubtreeLogNote")
     .def(pybind11::init<>())
@@ -391,7 +383,6 @@ node can have subnodes.)RAW")
     .value("Clock", sem::SubtreeLog::Kind::Clock)
     .value("State", sem::SubtreeLog::Kind::State)
     .value("Tag", sem::SubtreeLog::Kind::Tag)
-    .export_values()
     ;
   pybind11::class_<sem::Subtree, sem::SemId<sem::Subtree>, sem::Org>(m, "Subtree")
     .def_readwrite("loc", &sem::Subtree::loc, R"RAW(Document)RAW")
@@ -440,7 +431,6 @@ node can have subnodes.)RAW")
     .value("Deadline", sem::Subtree::Period::Kind::Deadline, R"RAW(Date of task completion. Must be a single time point)RAW")
     .value("Created", sem::Subtree::Period::Kind::Created, R"RAW(When the subtree was created)RAW")
     .value("Repeated", sem::Subtree::Period::Kind::Repeated, R"RAW(Last repeat time of the recurring tasks)RAW")
-    .export_values()
     ;
   pybind11::class_<sem::Subtree::Property>(m, "SubtreeProperty")
     .def(pybind11::init<>())
@@ -469,13 +459,11 @@ node can have subnodes.)RAW")
     .value("Override", sem::Subtree::Property::SetMode::Override)
     .value("Add", sem::Subtree::Property::SetMode::Add)
     .value("Subtract", sem::Subtree::Property::SetMode::Subtract)
-    .export_values()
     ;
   pybind11::enum_<sem::Subtree::Property::InheritanceMode>(m, "SubtreePropertyInheritanceMode")
     .value("ThisAndSub", sem::Subtree::Property::InheritanceMode::ThisAndSub)
     .value("OnlyThis", sem::Subtree::Property::InheritanceMode::OnlyThis)
     .value("OnlySub", sem::Subtree::Property::InheritanceMode::OnlySub)
-    .export_values()
     ;
   pybind11::class_<sem::Subtree::Property::Nonblocking>(m, "SubtreePropertyNonblocking")
     .def(pybind11::init<>())
@@ -522,7 +510,6 @@ node can have subnodes.)RAW")
     .value("Children", sem::Subtree::Property::Visibility::Level::Children)
     .value("Content", sem::Subtree::Property::Visibility::Level::Content)
     .value("All", sem::Subtree::Property::Visibility::Level::All)
-    .export_values()
     ;
   pybind11::class_<sem::Subtree::Property::ExportOptions>(m, "SubtreePropertyExportOptions")
     .def(pybind11::init<>())
@@ -560,7 +547,6 @@ node can have subnodes.)RAW")
     .value("Unnumbered", sem::Subtree::Property::Kind::Unnumbered)
     .value("Created", sem::Subtree::Property::Kind::Created)
     .value("Unknown", sem::Subtree::Property::Kind::Unknown)
-    .export_values()
     ;
   pybind11::class_<sem::LatexBody, sem::SemId<sem::LatexBody>, sem::Org>(m, "LatexBody")
     ;
@@ -646,7 +632,6 @@ node can have subnodes.)RAW")
     .value("None", sem::ListItem::Checkbox::None)
     .value("Done", sem::ListItem::Checkbox::Done)
     .value("Empty", sem::ListItem::Checkbox::Empty)
-    .export_values()
     ;
   pybind11::class_<sem::Link, sem::SemId<sem::Link>, sem::Org>(m, "Link")
     .def_readwrite("loc", &sem::Link::loc, R"RAW(Document)RAW")
@@ -685,7 +670,6 @@ node can have subnodes.)RAW")
     .value("Person", sem::Link::Kind::Person)
     .value("Footnote", sem::Link::Kind::Footnote)
     .value("File", sem::Link::Kind::File)
-    .export_values()
     ;
   pybind11::class_<sem::DocumentOptions, sem::SemId<sem::DocumentOptions>, sem::Org>(m, "DocumentOptions")
     .def_readwrite("loc", &sem::DocumentOptions::loc, R"RAW(Document)RAW")
@@ -729,13 +713,11 @@ node can have subnodes.)RAW")
   pybind11::enum_<sem::DocumentOptions::TocExportKind>(m, "DocumentOptionsTocExportKind")
     .value("DoExport", sem::DocumentOptions::TocExportKind::DoExport)
     .value("ExportFixed", sem::DocumentOptions::TocExportKind::ExportFixed)
-    .export_values()
     ;
   pybind11::enum_<sem::DocumentOptions::BrokenLinks>(m, "DocumentOptionsBrokenLinks")
     .value("Raise", sem::DocumentOptions::BrokenLinks::Raise)
     .value("Ignore", sem::DocumentOptions::BrokenLinks::Ignore)
     .value("Mark", sem::DocumentOptions::BrokenLinks::Mark)
-    .export_values()
     ;
   pybind11::enum_<sem::DocumentOptions::Visibility>(m, "DocumentOptionsVisibility")
     .value("Overview", sem::DocumentOptions::Visibility::Overview)
@@ -746,7 +728,6 @@ node can have subnodes.)RAW")
     .value("Show4Levels", sem::DocumentOptions::Visibility::Show4Levels)
     .value("Show5Levels", sem::DocumentOptions::Visibility::Show5Levels)
     .value("ShowEverything", sem::DocumentOptions::Visibility::ShowEverything)
-    .export_values()
     ;
   pybind11::class_<sem::Document, sem::SemId<sem::Document>, sem::Org>(m, "Document")
     .def_readwrite("loc", &sem::Document::loc, R"RAW(Document)RAW")
@@ -807,7 +788,6 @@ node can have subnodes.)RAW")
     .value("Export", sem::Include::Kind::Export)
     .value("Src", sem::Include::Kind::Src)
     .value("OrgDocument", sem::Include::Kind::OrgDocument)
-    .export_values()
     ;
   pybind11::class_<sem::DocumentGroup, sem::SemId<sem::DocumentGroup>, sem::Org>(m, "DocumentGroup")
     .def_readwrite("loc", &sem::DocumentGroup::loc, R"RAW(Document)RAW")
@@ -874,7 +854,6 @@ node can have subnodes.)RAW")
     .value("InheritanceMode_", OrgSpecName::InheritanceMode)
     .value("MainSetRule_", OrgSpecName::MainSetRule)
     .value("SubSetRule_", OrgSpecName::SubSetRule)
-    .export_values()
     ;
   pybind11::enum_<OrgNodeKind>(m, "OrgNodeKind")
     .value("None_", OrgNodeKind::None, R"RAW(Default valye for node - invalid state)RAW")
@@ -1064,7 +1043,6 @@ node can have subnodes.)RAW")
     .value("LogbookStateChange_", OrgNodeKind::LogbookStateChange, R"RAW(Annotation about change in the subtree todo state)RAW")
     .value("RadioTarget_", OrgNodeKind::RadioTarget, R"RAW(`<<<RADIO>>>`)RAW")
     .value("Target_", OrgNodeKind::Target, R"RAW(`<<TARGET>>`)RAW")
-    .export_values()
     ;
   pybind11::enum_<OrgBigIdentKind>(m, "OrgBigIdentKind")
     .value("None_", OrgBigIdentKind::None)
@@ -1149,7 +1127,6 @@ node can have subnodes.)RAW")
     .value("StructThen_", OrgBigIdentKind::StructThen)
     .value("StructElse_", OrgBigIdentKind::StructElse)
     .value("StructWhile_", OrgBigIdentKind::StructWhile)
-    .export_values()
     ;
   pybind11::enum_<OrgSemKind>(m, "OrgSemKind")
     .value("StmtList_", OrgSemKind::StmtList)
@@ -1206,7 +1183,6 @@ node can have subnodes.)RAW")
     .value("TextSeparator_", OrgSemKind::TextSeparator)
     .value("Include_", OrgSemKind::Include)
     .value("DocumentGroup_", OrgSemKind::DocumentGroup)
-    .export_values()
     ;
   pybind11::class_<OrgExporterJson>(m, "OrgExporterJson")
     .def(pybind11::init<>())
@@ -1371,7 +1347,6 @@ return state.)RAW")
     .value("QDateTime", LeafFieldType::QDateTime)
     .value("Str", LeafFieldType::Str)
     .value("Any", LeafFieldType::Any)
-    .export_values()
     ;
 }
 /* clang-format on */
